pyproject.toml
[project]
name = "bargain-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "pyjwt>=2.10.1",
    "pyngrok>=7.2.11",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
]
